Add filter toggler on small screen
